Update patch set 2
Patch Set 2:

We cannot keep this review open for long and I am not anticipating simple patch to this as it requires major changes. So, I am abandoning this commit and let us work on clean up of the code in bitbucket and then we can push the code as new commit once it is ready. Thanks.

Patch-set: 2
Reviewer: Gerrit User 3 <3@655300a2-7a56-4504-9153-881edd907bea>
Label: Verified=0