Update patch set 6
Patch Set 6:

FYI - I think this issue will bite you as well. In the AIO deploy I had to merge kong-migration and kong as there was a race condition between them, even with the wait for kong-database to be ready. There really is no reason for there to be a separate kong-migration container anyway. As it stands, there is no way to ensure that the migration is complete before kong tries to use the database, and in cases when that happens kong will fail.

Patch-set: 6