Update patch set 6
Patch Set 6:

Chris has a good point. Since the kong proxy is getting deployed already in production/dev/ist instances, is it using this exact set of tools/config? if not, this is for now superfluous and anyway needs to be reviewed in design per my comments. If this exact set of script/config *is* being used then it should be here, but in a folder designed to capture the tools/config used for non-AIO (multi-node) deploys - for that though I don't think we have yet settled how all those tools (beyond kong) should be managed thru this repo, so it's probably still preliminary to include this here. And I will soon implement multi-node support using the same AIO scripts, so even then this will be superfluous except where used in manual installs.

Patch-set: 6