Update patch set 2
Patch Set 2:

Sorry - wrong issue ID. I will correct that. The problem with the folder name is that it becomes part of the container name... so acumos_ is a more useful prefix for that. Remember back - I wanted to give the containers specific names but you pushed back on that. The result is that docker-compose creates the name based upon the folder name that contains the yml files. So "acumos" makes the most sense to me.

Patch-set: 2
Label: Code-Review=0