Update patch set 5
Patch Set 5:

OK, this is ready to merge. It's not fully working (still does nothing useful), but I have run out of patience with the current design team on this component, and will be taking it in a new direction. Merge this so I cap this effort and move forward with the common changes implemented as part of this patch.

Patch-set: 5
Reviewer: Gerrit User 93 <93@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0