Update patch set 1
Patch Set 1:

I don't have K8s in the VM on my laptop so I cannot test.

What components go together in a pod? How do you handle the Zuul reverse proxy that runs in the Portal-FE?  That has to route traffic to the BE and the DS docker containers.  Is that configured via names that K8s manages?

Patch-set: 1