Update patch set 1
Patch Set 1:

Chris: There is a pod per component. The deployment files are per component, and create a single instance per pod. The service files then create a frontend access port for that pod. Zuul is embedded in the portal-fe component/pod - I didn't change that. Routing to the backend components from Zull is handled by the variables and cluster-local DNS names as shown in the portal-fe deployment file.

Patch-set: 1