Update patch set 2
Patch Set 2:

Chris and Aimee - please let me know if you plan to test this before merging. I have tested it as noted for both clean and redeployment scenarios on both docker and k8s. I need to move on now to the actual upgrade part, which will be a bit more complex potentially, as upgrade involves in some cases other functional changes, and this redeployment capability adds a key missing feature.

Patch-set: 2