Update patch set 1
Patch Set 1:

(1 comment)

I will update this patch with similar templates for k8s, and test both docker and k8s. But I need to know what component assembly version is required, and if this change is in any way incompatible with earlier assemblies (this would need to be noted in the docs and script notes).

Patch-set: 1