Update patch set 1
Patch Set 1:

So we don't cause a merge issue on master, I suggest to abandon this patch. I will add the template as-is and fix it per Chris' comments into a new patch for assembly Acumos_1808171930, which covers both docker and k8s templates, as well as other changes planned for that assembly.

Patch-set: 1
Reviewer: Gerrit User 93 <93@655300a2-7a56-4504-9153-881edd907bea>
Label: Verified=0