Update patch set 10
Patch Set 10: Code-Review-1

Failed on a fresh Ubuntu 16.04.5.  The logs are empty.  I will re-run in a window so I have history.

The oneclick script still installs MariaDB, but I think you have switched to running a MariaDB docker image.  Is this correct, that you have taken a belt-and-suspenders approach?  If so, can you drop the install/remove of mariadb entirely?

Patch-set: 10
Label: Code-Review=-1