Update patch set 11
Patch Set 11:

OK, I used this as an option that by default is "in-docker": export ACUMOS_MARIADB=in-docker so even for docker-based deployment you are right MariaDB is installed under docker. The portal deployed fine for me but I am debugging an issue with onboarding under docker. Stay tuned.

Patch-set: 11