Update patch set 13
Patch Set 13:

On the 1st question - the debug logs should now be in /tmp/acumos/debug for both success and fail scenarios. It was a little tricky to get them saved while using sudo to dump them. On the mariadb question, mariadb server is deployed on the host or under k8s. Under k8s you still need the mariadb client on the host so that you can initialize the DB. Pretty soon (WIP) it will install the mariadb client on your workstation if you are deploying into a tenant account under k8s, since you as a tenant will not have privilege to install/config SW directly on the k8s cluster nodes.

Patch-set: 13