Update patch set 13
Patch Set 13:

> These messages are alarming, but do they really indicate an error?
 > 
 > WARNING: The ACUMOS_MARIADB_USER_PASSWORD variable is not set.    
 > Defaulting to a blank string.
 > ERROR: Couldn't connect to Docker daemon. You might need to install
 > Docker:

Those come from clean.sh which uses some brute-force methods, e.g. in this case trying to remove containers when they might not  be there to begin with, or the acumos-env.sh is not setup fully (thus the MARIADB warning), or docker isn't even running on the host. Could be more intelligent, but that intelligence sometimes misses corner cases that the blunt force approach addresses. But I will see if I can avoid those warnings, if you think it's really important.

Patch-set: 13