Update patch set 17
Patch Set 17:

> At finish the script directs me to https://ubuntu-cl778h:30885 but
 > I get connection refused on that port.  Checked Kong and it's not
 > listening on that port:
 > 
 > 4f55cd2ab2b        kong:0.11                                       
 >        "/docker-entrypoint.…"   14 minutes ago      Up 14 minutes  
 >     0.0.0.0:30080->8000/tcp, 0.0.0.0:30081->8001/tcp,
 > 0.0.0.0:30443->8443/tcp, 0.0.0.0:30444->8444/tcp  
 > acumos_kong-service_1_258e1920df2e
 > 
 > 
 > clean script does this:
 > 
 > ./clean.sh: line 42: kubectl: command not found
 > 
 > 
 > Still cannot 100% verify due to apparent DNS block on 8.8.8.8 used
 > by logstash docker image build process.

The incorrect port was output - just a logic error in the user hint. Fixed in the next patch. The portal is accessible on port 30443
Also will fix the attempts to use kubectl (innocuous) when it does not exist in the path.
Not sure what you are referring to re logstash. It builds and deploys fine for me. Maybe you are behind a blocking firewall?

Patch-set: 17
Label: Code-Review=0