Update patch set 23
Patch Set 23:

> I would like to provide a log.  The script tells me to look in
 > /home/clott/acumos/logs but that directory is empty.  Was I
 > supposed to add some argument to force creation of a log?

I'm still working on a consistent approach to saving logs. It's possible based upon the nature of the error that no log was saved. For now I will remove that part of the script so it's not giving you the wrong impression that is has saved logs.

To log the script, add ' 2>&1 | tee <logname>' to the end of the command.

Patch-set: 23