Update patch set 7
Patch Set 7:

The cleanup is now part of the redeployment process. You do not need a separate cleanup step. If you need to clean the parts that the admin user prepares, then start with that script. This is relative to the type of deployment you are doing; but in all cases the safest thing to do in a clean redeployment case is to start with the prep scripts (setup_prereqs.sh or acumos_k8s_prep.sh).

Patch-set: 7