Update patch set 7
Patch Set 7:

> Please tell me, is there a script to stop all the running
 > containers and possibly remove the images that were pulled?  Lower
 > priority, is there a way to undo all the prereqs that were set up?

There isn't a specific script to stop all the running containers and remove the images, or to undo the host customizations of the prereq scripts. The assumption here is that a host is dedicated to this service, and as need be will be rebuilt for redeployment. Cleaning and redeploying will be supported and within the limits of our ability to handle the plethora of variations that result from deploying into a cleaned but still dirty environment, it will be reliable. But cleaning itself has been integrated into the deployment tools. I can break that out again if really needed.

Patch-set: 7
Label: Code-Review=0