Update patch set 11
Patch Set 11:

Ran setup_prereqs for docker, and oneclick_deploy successfully.  Visited the page. Unfortunately the behavior I see is that the home screen pups up a dialog "Please sign in to the application" repeatedly. I click OK to dismiss the dialog, it immediately reappears.  I don't think this is a behavior that AIO has caused or can control, but the user experience is that logging in is impossible.

Patch-set: 11