Update patch set 19
Patch Set 19:

It does that to determine if the platform was installed under k8s, or if at least the kubectl command is installed so it can try to cleanup any platform in the acumos namespace. It's a fairly weak but effective approach to cleanup in case acumos_env.sh does not accurately reflect the as-installed options.

Patch-set: 19