Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> why are you copying this code from their respective repos?

Hi, the helm charts for license management works well when we deploy them independently. But when we try to deploy them into same namespace, there is a conflict about the acumos-registry
secret. Since the Acumos docker repo doesn't need the pull image secret, I took that part out. Also I have a concern about the init db script in lum-helm, the reference will not work if we push the chats to a helm repo.

But we are not sure that code change will affect the current Acumos deployment. So we decided to save the charts in system-integration repo. Once we test the charts we would contribute that back to license management project.

Patch-set: 2