Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-1
> > 
> > why are you copying this code from their respective repos?
> 
> Hi, the helm charts for license management works well when we deploy them independently. But when we try to deploy them into same namespace, there is a conflict about the acumos-registry
> secret. Since the Acumos docker repo doesn't need the pull image secret, I took that part out. Also I have a concern about the init db script in lum-helm, the reference will not work if we push the chats to a helm repo.
> 
> But we are not sure that code change will affect the current Acumos deployment. So we decided to save the charts in system-integration repo. Once we test the charts we would contribute that back to license management project.

It is difficult to determine what you copied and what you changed do you have a list of files you changed to fix the items above? Ok to be clear this is just a test case and not planned to merge into master. Keeping -1 we should use subcharts or use helmfile to pull in multiple helm charts. We do not want to maintain multiple copies of helm + sql files. I have tested the charts with different namespaces or the same as acumos and didn't face issues other than the docker registry secret which had to be created only if using a different namespace other than acumos. I was using helm 3 without tiller for my testing. What version of helm are you using/testing with? Is there any documentation of running helm install you are trying to get working?

Patch-set: 2