Update patch set 3
Patch Set 3:

(1 comment)

Hi Vaibhav,

The helm charts for MLWB looks much better now. But still have a one tiny issue.

In some component deployment.yaml, the service name was assign to the container name. But others, the container name was a constant string. 

The reason why, we parameterize the podName and svcName, is we want to avoid conflict of kubernetes object name, if we deploy same component into same namespace. But that will not happen on container name, so in other Acumos helm charts we just define container name as a constant string.

Patch-set: 3
Reviewer: Gerrit User 381 <381@655300a2-7a56-4504-9153-881edd907bea>