Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > > Patch Set 1:
> > > 
> > > Thanks for +1, I just notice that we have an "http_proxy" section just few line below my code change
> > > 
> > > "http_proxy": "http://{{ .Values.global.acumosProxy }}.{{ .Values.global.namespace }}:{{ .Values.global.acumosProxyPort }}",
> > > 
> > > this "http_proxy" has exactly the same value than the "proxyPrefix" I added .... so perhaps we should remove this http_proxy ?
> > 
> > Philippe Does this mean that microservice generation is not looking for "http_proxy"?
> 
> I don't think so, In my point of view there was a confusing between http-proxy and the proxyPrefix, in most of the onboarding case we don't need proxy it is only for onboarding pre-dockerized model. Let me check with Guy

I made several test locally and the result is that I am wrong we must keep this http_proxy in the deployment yaml file,whithout it acumos cannot fecth external ressources to create docker images. But 
I am still convince that what I added in "imagetag" is not the right value because it not solve the defect ACUMOS-4276.

Patch-set: 1